CDP #142 - Moving NavItem, NavLink, NavSearch components to "components"
